% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/designPrior.R
\name{density.designPrior}
\alias{density.designPrior}
\title{Density method for design prior object}
\usage{
\method{density}{designPrior}(x, ...)
}
\arguments{
\item{x}{A designPrior object}

\item{...}{Other arguments}
}
\description{
Density method for design prior object
}
\examples{
dp <- designPrior(to = 2.3123, so = 0.1, mu = 1.1, tau = 0.2)
f <- density(dp)
tseq <- seq(1, 3, 0.01)
plot(tseq, f(theta = tseq), type = "l", xlab = "theta", ylab = "Design prior density")
}
