% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/designPrior.R
\name{plot.designPrior}
\alias{plot.designPrior}
\title{Plot method for design prior object}
\usage{
\method{plot}{designPrior}(x, ...)
}
\arguments{
\item{x}{A designPrior object}

\item{...}{Other arguments}
}
\description{
Plot method for design prior object
}
\examples{
dp <- designPrior(to = 2.3123, so = 0.1, mu = 1.1, tau = 0.2)
plot(dp)
plot(dp, xlim = c(0, 5), length.out = 500)
}
