% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssdBFr.R
\name{porsBFr}
\alias{porsBFr}
\title{Probability of replication success based on replication Bayes factor}
\usage{
porsBFr(level, dprior, sr)
}
\arguments{
\item{level}{Bayes factor level below which replication success is achieved}

\item{dprior}{Design prior object}

\item{sr}{Replication standard error}
}
\value{
The probability to achieve replication success
}
\description{
This function computes the probability to achieve replication
    success based on the replication Bayes factor
}
\examples{
## specify design prior
to1 <- 0.2
so1 <- 0.05
dprior <- designPrior(to = to1, so = so1, tau = 0.03)
porsBFr(level = 1/10, dprior = dprior, sr = c(0.05, 0.04))

}
\references{
Pawel, S., Consonni, G., and Held, L. (2022). Bayesian approaches to
designing replication studies. arXiv preprint.
\doi{10.48550/arXiv.XXXX.XXXXX}
}
\author{
Samuel Pawel
}
