% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssdBFs.R
\name{porsBFs}
\alias{porsBFs}
\title{Probability of replication success based on the sceptical Bayes factor}
\usage{
porsBFs(level, dprior, sr)
}
\arguments{
\item{level}{Threshold for the sceptical Bayes factor below which replication
success is achieved}

\item{dprior}{Design prior object}

\item{sr}{Replication standard error}
}
\value{
The probability to achieve replication success
}
\description{
This function computes the probability to achieve replication
    success based on the Bayes factor
}
\examples{
## specify design prior
to1 <- 0.2
so1 <- 0.05
dprior <- designPrior(to = to1, so = so1)
porsBFs(level = 1/38.7, dprior = dprior, sr = 0.05)

}
\references{
Pawel, S., Consonni, G., and Held, L. (2022). Bayesian approaches to
designing replication studies. arXiv preprint.
\doi{10.48550/arXiv.XXXX.XXXXX}
}
\author{
Samuel Pawel
}
