% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/designPrior.R
\name{print.designPrior}
\alias{print.designPrior}
\title{Print method for design prior object}
\usage{
\method{print}{designPrior}(x, ...)
}
\arguments{
\item{x}{A designPrior object}

\item{...}{Other arguments}
}
\description{
Print method for design prior object
}
\examples{
dp <- designPrior(to = 0.5, so = 0.05, sp = 0.2, tau = 0.1)
print(dp)
}
