% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssdRS.R
\name{print.ssdRS}
\alias{print.ssdRS}
\title{Print method for ssdRS object}
\usage{
\method{print}{ssdRS}(x, ...)
}
\arguments{
\item{x}{An ssdRS object}

\item{...}{Other arguments}
}
\description{
Print method for ssdRS object
}
\examples{
## specify design prior
to1 <- 2
so1 <- 1
dprior <- designPrior(to = to1, so = so1)

## compute required standard error for significance at one-sided 2.5\%
sregionfunSig <- function(sr, alpha = 0.025) {
    successRegion(intervals = cbind(stats::qnorm(p = 1- alpha)*sr, Inf))
}
ssd1 <- ssdRS(sregionfun = sregionfunSig, dprior = dprior, power = 0.8)
print(ssd1)
}
