% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/successRegion.R
\name{print.successRegion}
\alias{print.successRegion}
\title{Print method for successRegion object}
\usage{
\method{print}{successRegion}(x, ...)
}
\arguments{
\item{x}{A successRegion object}

\item{...}{Other arguments}
}
\description{
Print method for successRegion object
}
\examples{
## success region for two-sided significance test
successRegion(intervals = rbind(c(1.96, Inf), c(-Inf, -1.96)))
## success region for one-sided significance test
successRegion(intervals = rbind(c(1.96, Inf)))
}
