% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssdRS.R
\name{ssdRS}
\alias{ssdRS}
\title{Sample size determination for replication success}
\usage{
ssdRS(
  sregionfun,
  dprior,
  power,
  searchinterval = c(.Machine$double.eps^0.5, 100),
  ...
)
}
\arguments{
\item{sregionfun}{Function that returns the success region for replication
effect estimate as a function of the replication standard error}

\item{dprior}{Design prior object}

\item{power}{Desired probability of replication success}

\item{searchinterval}{Search interval for standard errors}

\item{...}{other arguments for uniroot}
}
\value{
An ssdRS object containing the design prior, the recomputed power,
the required replication standard error sr, and the relative sample size
c = so^2/sr^2 = nr/no
}
\description{
This function computes the standard error required to achieve
replication success with a certain probability and based on a certain
type of success region.
}
\details{
The probability of replication success .....
}
\references{
Pawel, S., Consonni, G., and Held, L. (2022). Bayesian approaches to
designing replication studies. arXiv preprint.
\doi{10.48550/arXiv.XXXX.XXXXX}
}
\author{
Samuel Pawel
}
