% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/designPrior.R
\name{designPrior}
\alias{designPrior}
\title{Design prior for effect size}
\usage{
designPrior(
  to,
  so,
  mu = 0,
  sp = Inf,
  tau = 0,
  g = (sp^2 + tau^2)/so^2,
  h = tau^2/so^2,
  type = c(NA, "conditional", "predictive", "EB", "simple")
)
}
\arguments{
\item{to}{Effect estimate from original study \eqn{\hat{\theta}_o}{to}}

\item{so}{Standard error of effect estimate from original study
\eqn{\sigma_o}{so}}

\item{mu}{The prior mean \eqn{\mu_{\theta}}{mu}. The default is zero.}

\item{sp}{The prior standard deviation \eqn{\sigma_{\theta}}{sp}. The default
is infinity (an improper uniform prior).}

\item{tau}{The prior heterogeneity standard deviation \eqn{\tau}{tau}. The
default is zero (no heterogeneity).}

\item{g}{The relative prior variance \eqn{g = (\sigma_{\theta}^2 +
\tau^2)/\sigma_o^2}{g = (sp^2 + tau^2)/so^2} (alternative parametrization
of prior standard deviation \eqn{\sigma_{\theta}}{sp})}

\item{h}{The relative prior heterogeneity variance \eqn{h =
\tau^2/\sigma_o^2}{h = tau^2/so^2} (alternative parametrization of prior
heterogeneity standard deviation \eqn{\tau}{tau})}

\item{type}{Shortcut for special parameter combinations. The available
options are NA, "conditional", "predictive", "EB", and "simple" (see
below for details). Defaults to NA.}
}
\value{
A designPrior object
}
\description{
Creates a design prior object.
}
\examples{
designPrior(to = 1.1, so = 1)
}
\references{
Pawel, S., Consonni, G., and Held, L. (2022). Bayesian approaches to
designing replication studies. arXiv preprint.
\doi{10.48550/arXiv.XXXX.XXXXX}
}
\seealso{
\code{\link{probRS}}
}
\author{
Samuel Pawel
}
