% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/designPrior.R
\name{plot.designPrior}
\alias{plot.designPrior}
\title{Plot method for class \code{"designPrior"}}
\usage{
\method{plot}{designPrior}(x, ...)
}
\arguments{
\item{x}{Object of class \code{"designPrior"}}

\item{...}{Other arguments passed to \code{plot}}
}
\value{
Plots the density of the design prior
}
\description{
Plot method for class \code{"designPrior"}
}
\examples{
dp <- designPrior(to = 2.3123, so = 0.1, mu = 1.1, tau = 0.2)
plot(dp)
plot(dp, xlim = c(0, 5), length.out = 500)
}
\author{
Samuel Pawel
}
