% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pors.R
\name{pors}
\alias{pors}
\title{Compute probability of replication success}
\usage{
pors(sregion, dprior, sr)
}
\arguments{
\item{sregion}{Success region for replication effect estimate}

\item{dprior}{Design prior object}

\item{sr}{Standard error of Replication effect estimate}
}
\value{
The probability of replication success
}
\description{
This function computes the probabiliy of replication success
    based on a success region for the replication effect estimate, a
    design prior, and a replication standard error.
}
\examples{
dprior <- designPrior(to = 1.1, so = 1)
sregion <- successRegion(intervals = cbind(1.96, Inf))
pors(sregion = sregion, dprior = dprior, sr = 1)

}
\references{
Pawel, S., Consonni, G., and Held, L. (2022). Bayesian approaches to
designing replication studies. arXiv preprint.
\doi{10.48550/arXiv.XXXX.XXXXX}
}
\author{
Samuel Pawel
}
