% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssd.R
\name{ssd}
\alias{ssd}
\title{Sample size determination for replication success}
\usage{
ssd(sregionfun, dprior, power, searchInt = c(.Machine$double.eps^0.5, 2), ...)
}
\arguments{
\item{sregionfun}{Function that returns the success region for replication
effect estimate as a function of the replication standard error}

\item{dprior}{Design prior object}

\item{power}{Desired probability of replication success}

\item{searchInt}{Search interval for standard errors}

\item{...}{other arguments for uniroot}
}
\value{
An ssdRS object containing the design prior, the recomputed power,
the required replication standard error sr, and the relative sample size
c = so^2/sr^2 = nr/no
}
\description{
This function computes the standard error required to achieve
replication success with a certain probability and based on a certain
type of success region.
}
\details{
The probability of replication success .....
}
\examples{
## specify design prior
to1 <- 2
so1 <- 1
dprior <- designPrior(to = to1, so = so1)

## compute required standard error for significance at one-sided 2.5\%
sregionfunSig <- function(sr, alpha = 0.025) {
    successRegion(intervals = cbind(stats::qnorm(p = 1- alpha)*sr, Inf))
}
ssd(sregionfun = sregionfunSig, dprior = dprior, power = 0.8)

}
\references{
Pawel, S., Consonni, G., and Held, L. (2022). Bayesian approaches to
designing replication studies. arXiv preprint.
\doi{10.48550/arXiv.XXXX.XXXXX}
}
\author{
Samuel Pawel
}
