% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssdBFr.R
\name{ssdBFr}
\alias{ssdBFr}
\title{Sample size determination for replication success based on
    replication Bayes factor}
\usage{
ssdBFr(level, dprior, power, searchInt = c(.Machine$double.eps^0.5, 2))
}
\arguments{
\item{level}{Bayes factor level below which replication success is achieved}

\item{dprior}{Design prior object}

\item{power}{Desired probability of replication success}

\item{searchInt}{Interval for numerical search over replication standard
errors}
}
\value{
An ssdRS object containing the design prior, the recomputed power,
    the required replication standard error sr, and the relative sample size
    c = so^2/sr^2 = nr/no
}
\description{
This function computes the standard error required to achieve
    replication success with a certain probability and based on the
    replication Bayes factor under normality
}
\examples{
## specify design prior
to1 <- 0.2
so1 <- 0.05
dprior <- designPrior(to = to1, so = so1, tau = 0.03)
ssdBFr(level = 1/10, dprior = dprior, power = 0.8)

}
\references{
Pawel, S., Consonni, G., and Held, L. (2022). Bayesian approaches to
designing replication studies. arXiv preprint.
\doi{10.48550/arXiv.XXXX.XXXXX}

Verhagen, J. and Wagenmakers, E. J. (2014). Bayesian tests to quantify the result
of a replication attempt. Journal of Experimental Psychology: General,
145:1457-1475. \doi{10.1037/a0036731}

Ly, A., Etz, A., Marsman, M., & Wagenmakers, E.-J. (2018). Replication Bayes
factors from evidence updating. Behavior Research Methods, 51(6), 2498-2508.
\doi{10.3758/s13428-018-1092-x}
}
\author{
Samuel Pawel
}
