% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/successRegion.R
\name{successRegion}
\alias{successRegion}
\title{Success region for replication effect estimate}
\usage{
successRegion(intervals)
}
\arguments{
\item{intervals}{A 2xN matrix containing N disjoint intervals, the first
entry containing the lower and the second entry containing the upper
limit}
}
\value{
A successRegion object
}
\description{
Creates a success region object
}
\examples{
successRegion(intervals = rbind(c(1.96, Inf), c(-Inf, -1.96)))
successRegion(intervals = cbind(1.96, Inf))

}
\references{
Pawel, S., Consonni, G., and Held, L. (2022). Bayesian approaches to
designing replication studies. arXiv preprint.
\doi{10.48550/arXiv.XXXX.XXXXX}
}
\seealso{
\code{\link{probRS}}
}
\author{
Samuel Pawel
}
